[app] Clean up examples for getting started.
