+++ app/src/androidTest/java/com/mapbox/maps/testapp/integration/surface/SurfaceViewOrientationChangeTest.kt
@@ -3,7 +3,7 @@ package com.mapbox.maps.testapp.integration.surface
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.mapbox.maps.testapp.examples.SimpleMapActivity
+import com.mapbox.maps.testapp.examples.gettingstarted.SimpleMapActivity
 import com.mapbox.maps.testapp.integration.BaseIntegrationTest
 import org.junit.Rule
 import org.junit.Test
