+++ app/src/androidTest/java/com/mapbox/maps/testapp/integration/surface/SurfaceViewReopenTest.kt
@@ -3,8 +3,8 @@ package com.mapbox.maps.testapp.integration.surface
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.mapbox.maps.testapp.examples.SimpleMapActivity
-import com.mapbox.maps.testapp.examples.TextureViewActivity
+import com.mapbox.maps.testapp.examples.gettingstarted.SimpleMapActivity
+import com.mapbox.maps.testapp.examples.gettingstarted.TextureViewActivity
 import com.mapbox.maps.testapp.integration.BaseIntegrationTest
 import com.mapbox.maps.testapp.integration.launchActivity
 import org.junit.Rule
