+++ app/src/androidTest/java/com/mapbox/maps/testapp/integration/texture/TextureViewOrientationChangeTest.kt
@@ -3,7 +3,7 @@ package com.mapbox.maps.testapp.integration.texture
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.mapbox.maps.testapp.examples.TextureViewActivity
+import com.mapbox.maps.testapp.examples.gettingstarted.TextureViewActivity
 import com.mapbox.maps.testapp.integration.BaseIntegrationTest
 import org.junit.Rule
 import org.junit.Test
