+++ app/src/androidTest/java/com/mapbox/maps/testapp/integration/texture/TextureViewReopenTest.kt
@@ -3,7 +3,7 @@ package com.mapbox.maps.testapp.integration.texture
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.mapbox.maps.testapp.examples.TextureViewActivity
+import com.mapbox.maps.testapp.examples.gettingstarted.TextureViewActivity
 import com.mapbox.maps.testapp.integration.BaseIntegrationTest
 import com.mapbox.maps.testapp.integration.launchActivity
 import org.junit.Rule
@@ -14,7 +14,9 @@ import org.junit.runner.RunWith
 class TextureViewReopenTest : BaseIntegrationTest() {
 
   @get:Rule
-  var activityRule: ActivityScenarioRule<TextureViewActivity> = ActivityScenarioRule(TextureViewActivity::class.java)
+  var activityRule: ActivityScenarioRule<TextureViewActivity> = ActivityScenarioRule(
+    TextureViewActivity::class.java
+  )
 
   @Test
   @LargeTest
