+++ app/src/main/AndroidManifest.xml
@@ -31,25 +31,25 @@
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" />
         <activity
-            android:name=".examples.SimpleMapActivity"
+            android:name=".examples.gettingstarted.SimpleMapActivity"
             android:description="@string/description_simple_map"
             android:exported="true"
             android:label="@string/activity_simple_map">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
         </activity>
         <activity
-            android:name=".examples.MapboxStudioStyleActivity"
+            android:name=".examples.gettingstarted.MapboxStudioStyleActivity"
             android:description="@string/description_studio_style_map"
             android:exported="true"
             android:label="@string/activity_studio_style">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
@@ -61,7 +61,7 @@
             android:label="@string/activity_overlay">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
@@ -248,13 +248,13 @@
                 android:value=".ExampleOverviewActivity" />
         </activity>
         <activity
-            android:name=".examples.DebugModeActivity"
+            android:name=".examples.gettingstarted.DebugModeActivity"
             android:description="@string/description_debug_mode"
             android:exported="true"
             android:label="@string/activity_debug_mode">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
@@ -428,13 +428,13 @@
                 android:value=".ExampleOverviewActivity" />
         </activity>
         <activity
-            android:name=".examples.TextureViewActivity"
+            android:name=".examples.gettingstarted.TextureViewActivity"
             android:description="@string/description_texture_view"
             android:exported="true"
             android:label="@string/activity_texture_view">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
@@ -674,7 +674,7 @@
             android:label="@string/activity_basic_add_marker_title">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
@@ -686,7 +686,7 @@
             android:label="@string/activity_basic_add_one_marker_title">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
@@ -974,7 +974,7 @@
             android:label="@string/activity_map_view_customization">
             <meta-data
                 android:name="@string/category"
-                android:value="@string/category_basic" />
+                android:value="@string/getting_started" />
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ExampleOverviewActivity" />
