+++ app/src/main/java/com/mapbox/maps/testapp/examples/gettingstarted/MapboxStudioStyleActivity.kt
@@ -1,26 +1,19 @@
-package com.mapbox.maps.testapp.examples
+package com.mapbox.maps.testapp.examples.gettingstarted
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.mapbox.maps.MapboxMap
-import com.mapbox.maps.Style
 import com.mapbox.maps.testapp.R
-import kotlinx.android.synthetic.main.activity_texture_view.*
+import kotlinx.android.synthetic.main.activity_simple_map.*
 
 /**
- * Example of displaying a map using TextureView as render surface.
+ * Example of displaying a custom Mapbox-hosted style.
  */
-class TextureViewActivity : AppCompatActivity() {
-
-  private lateinit var mapboxMap: MapboxMap
+class MapboxStudioStyleActivity : AppCompatActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    setContentView(R.layout.activity_texture_view)
-
-    mapboxMap = mapView.getMapboxMap()
-
-    mapboxMap.loadStyleUri(Style.MAPBOX_STREETS)
+    setContentView(R.layout.activity_style_mapbox_studio)
+    mapView.getMapboxMap().loadStyleUri(STUDIO_STYLE_URL)
   }
 
   override fun onStart() {
@@ -41,5 +34,9 @@ class TextureViewActivity : AppCompatActivity() {
   override fun onDestroy() {
     super.onDestroy()
     mapView.onDestroy()
+  }
+
+  companion object {
+    private const val STUDIO_STYLE_URL = "mapbox://styles/mapbox/cj3kbeqzo00022smj7akz3o1e"
   }
 }
