+++ app/src/main/java/com/mapbox/maps/testapp/examples/gettingstarted/SimpleMapActivity.kt
@@ -1,4 +1,4 @@
-package com.mapbox.maps.testapp.examples
+package com.mapbox.maps.testapp.examples.gettingstarted
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
