+++ app/src/main/res/layout/activity_cache_manager.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.SimpleMapActivity">
+    tools:context=".examples.gettingstarted.SimpleMapActivity">
 
     <TextView
         android:id="@+id/resultView"
