+++ app/src/main/res/values/example_descriptions.xml
@@ -18,7 +18,7 @@
     <string name="description_map_fragment">Use a MapView inside a map fragment and fragment backstack</string>
     <string name="description_map_viewpager">Use a MapView inside a ViewPager</string>
     <string name="description_map_multi">Using multiple map fragments in one Activity</string>
-    <string name="description_debug_mode">Use variations of debug modes of the Map</string>
+    <string name="description_debug_mode">Use variations of debug modes with a map view</string>
     <string name="description_snapshotter">Generate a static map image</string>
     <string name="description_snapshotter_dds">Use data driven styling with static map image generation</string>
     <string name="description_snapshotter_local">Use an local style with static map image generation</string>
